🚚 Move config classes to separate folder
